Add create chat api
Create a chat record in the chats collection in mongodb by using the
createChat function under ".../controllers/chat.js". User now has
listOfChats field that includes chat they are a part of. Duplicate chat
creation requests are ignored. A request is considered duplicate if two
users already are a part of another chat.